Newmaster (#29)
* Tweak ldma logger a bit.

* 8-bit radio channel.

* added flash partition support

* Add missing header include.

* Added mutex delete function.

* __ASM not standard, use __asm__ ?

* Add EUI64 data type and support functions.

* Increase acq time for internal temp.

* Support cmsis.

* Monitor and fix polarity errors.

* added spi flash sleep support

* Add DMADRV support to LDMA logger.

* Add ADC sharing components, implement in mcu temp reader.

* Fix typo in ADC list.

* retry support

* BARR spi_flash.

* Make mcu dataflash usable with fs.

* Reset USART before configuring.

* Update radio_basic with PA/antenna config.

* Platform mutex is universal, move silabs->common.

* RF2G4 not RF2G2.

* serial_activemessage moved to mist-comm.

* Convert BME wrapper for latest Bosch driver.

* Add basic logger for pre-kernel logging.

* Basic filesystem and logger setup scenarios.

* Some common compat files.

* Ident parameters.

* Add many time functions.

* Add flash sleep tracking option.

* Convert sleep times to 64bit.

* Bugs and tweaks.

* Platform I2C resource manager.

* Basic sleep stats helper.

* Basic radio duty-cycling component.

* Specify 3K stack size for radio.

* Panic on unsuccessful wakeup.

* Wake dataflash from deep sleep

* Warn when sys_panic will while(1).

* Handle a previously unhandled RAIL error.

* Add APIs for coordinates.

* Added function to retrieve platform eui.

* Added mock platform eui.

* Added check for all series.

* Add node_lifetime header.

* Added widget to monitor threads.

* Changed ldma thread stack size to fixed number.

* Fix typo.

* Use new lll API.

* Fix erase and misaligned short data bugs. Disable interrupts just in case.

* Fix a bit of whitespace.

* Print both words and bytes.

* Add heapstats function.

+++ include/silabs/lptsleep.h
@@ -56,6 +56,6 @@ void vLowPowerSleepTimerSetFreq(uint32_t freq);
  * Get the time spent in sleep - milliseconds.
  * @return Time spent sleeping.
  */
-uint32_t ulLowPowerSleepTime();
+uint64_t ulLowPowerSleepTime();
 
 #endif//LPTSLEEP_H_
